remove all policy-related stuffs
