--- src/abe_policy/policy/parser.rs
@@ -1,292 +1,0 @@
-use std::collections::HashMap;
-
-use serde_json::Value;
-
-use crate::{
-    abe_policy::{AttributeParameters, Dimension, DimensionBuilder, EncryptionHint},
-    Error,
-};
-
-use super::{
-    policy_v2::PolicyV2,
-    policy_versions::{LegacyPolicy, PolicyV1},
-    PolicyVersion,
-};
-
-impl PolicyV2 {
-    /// Converts the given bytes into a Policy, supports legacy policy versions.
-    pub fn parse_and_convert(bytes: &[u8]) -> Result<Self, Error> {
-        let json_policy: Value =
-            serde_json::from_slice(bytes).map_err(Error::DeserializationError)?;
-
-        if let Some(policy_version) = json_policy.get("version") {
-            match serde_json::from_value::<PolicyVersion>(policy_version.clone()) {
-                Ok(PolicyVersion::V1) => Ok(serde_json::from_slice::<PolicyV1>(bytes)
-                    .map_err(Error::DeserializationError)?
-                    .into()),
-                Ok(PolicyVersion::V2) => {
-                    serde_json::from_value::<Self>(json_policy).map_err(Error::DeserializationError)
-                }
-                Ok(_) => todo!(),
-                Err(e) => Err(Error::DeserializationError(e)),
-            }
-        } else {
-            // Legacy Policy
-            Ok(serde_json::from_slice::<LegacyPolicy>(bytes)
-                .map_err(Error::DeserializationError)?
-                .into())
-        }
-    }
-}
-
-impl TryFrom<&[u8]> for PolicyV2 {
-    type Error = Error;
-
-    fn try_from(bytes: &[u8]) -> Result<Self, Self::Error> {
-        Self::parse_and_convert(bytes)
-    }
-}
-
-impl TryFrom<&PolicyV2> for Vec<u8> {
-    type Error = Error;
-
-    fn try_from(policy: &PolicyV2) -> Result<Self, Self::Error> {
-        serde_json::to_vec(policy).map_err(Self::Error::DeserializationError)
-    }
-}
-
-impl TryFrom<HashMap<String, Vec<String>>> for PolicyV2 {
-    type Error = Error;
-
-    /// Create a policy object from policy specifications
-    ///
-    /// The policy specifications format is as follow:
-    /// ```text
-    /// policy: \{ dimension* \}
-    /// dimension: "dimension_name[ hierarchical_modifier ]": \[ attribute* \]
-    /// dimension_name: String
-    /// hierarchical_modifier : "::<"
-    /// attribute: "attribute_name[ hybridized_modifier ]"
-    /// attribute_name: String
-    /// hybridized_modifier: "::+"
-    /// ```
-    ///
-    /// Example:
-    /// ```json
-    ///     {
-    ///        "Security Level::<": [
-    ///            "Protected",
-    ///            "Confidential",
-    ///            "Top Secret::+"
-    ///        ],
-    ///        "Department": [
-    ///            "R&D",
-    ///            "HR",
-    ///            "MKG",
-    ///            "FIN"
-    ///        ]
-    ///    }
-    /// ```
-    fn try_from(value: HashMap<String, Vec<String>>) -> Result<Self, Self::Error> {
-        let mut policy = Self::new();
-
-        for (dimension, attributes) in &value {
-            // Split the dimension into name and hierarchy flag
-            let (dim_name, hierarchical) = match dimension.split_once("::") {
-                Some((name, specs)) => {
-                    // If the dimension contains the hierarchy flag, parse it
-                    let hierarchical = match specs {
-                        "<" => true,
-                        x => {
-                            return Err(Error::ConversionFailed(format!(
-                                "invalid specification '{x}' for dimension '{name}'"
-                            )));
-                        }
-                    };
-                    (name, hierarchical)
-                }
-                // If there is no hierarchy flag, assume the dimension is non-hierarchical
-                None => (dimension.as_str(), false),
-            };
-
-            let mut attributes_properties: Vec<(&str, EncryptionHint)> =
-                Vec::with_capacity(attributes.len());
-
-            // Parse each attribute and its encryption hint
-            for att in attributes {
-                let (att_name, encryption_hint) = match att.split_once("::") {
-                    Some((name, specs)) => {
-                        let encryption_hint = match specs {
-                            "+" => EncryptionHint::Hybridized,
-                            x => {
-                                return Err(Error::ConversionFailed(format!(
-                                    "invalid specification '{x}' for attribute \
-                                     '{dim_name}::{name}'"
-                                )));
-                            }
-                        };
-                        (name, encryption_hint)
-                    }
-                    // If there is no encryption hint, assume the attribute is non-hybridized
-                    None => (att.as_str(), EncryptionHint::Classic),
-                };
-                attributes_properties.push((att_name, encryption_hint));
-            }
-
-            // Add the dimension to the policy
-            policy.add_dimension(DimensionBuilder::new(
-                dim_name,
-                attributes_properties,
-                hierarchical,
-            ))?;
-        }
-        Ok(policy)
-    }
-}
-
-impl TryFrom<PolicyV2> for HashMap<String, Vec<String>> {
-    type Error = Error;
-
-    fn try_from(policy: PolicyV2) -> Result<Self, Self::Error> {
-        fn convert_attribute(attribute: (String, AttributeParameters)) -> String {
-            let (name, params) = attribute;
-            match params.get_encryption_hint() {
-                EncryptionHint::Hybridized => name + "::+",
-                EncryptionHint::Classic => name,
-            }
-        }
-        let mut result: Self = Self::with_capacity(policy.dimensions.len());
-
-        for (dim_name, dimension) in policy.dimensions {
-            let (dim_full_name, attributes_list) = match dimension {
-                Dimension::Unordered(attributes) => {
-                    let attributes_list: Vec<String> =
-                        attributes.into_iter().map(convert_attribute).collect();
-                    (dim_name, attributes_list)
-                }
-                Dimension::Ordered(attributes) => {
-                    let dim_name = dim_name + "::<";
-                    let attributes_list: Vec<String> =
-                        attributes.into_iter().map(convert_attribute).collect();
-                    (dim_name, attributes_list)
-                }
-            };
-            result.insert(dim_full_name, attributes_list);
-        }
-        Ok(result)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use std::collections::HashMap;
-
-    use crate::{
-        abe_policy::{Dimension, DimensionBuilder, EncryptionHint, Policy, QualifiedAttribute},
-        Error,
-    };
-
-    #[test]
-    pub fn test_parse_policy_from_bytes() -> Result<(), Error> {
-        let mut policy = Policy::new();
-        policy.add_dimension(DimensionBuilder::new(
-            "Security Level",
-            vec![
-                ("Protected", EncryptionHint::Classic),
-                ("Confidential", EncryptionHint::Classic),
-                ("Top Secret", EncryptionHint::Hybridized),
-            ],
-            true,
-        ))?;
-        policy.add_dimension(DimensionBuilder::new(
-            "Department",
-            vec![
-                ("R&D", EncryptionHint::Classic),
-                ("HR", EncryptionHint::Classic),
-                ("MKG", EncryptionHint::Classic),
-                ("FIN", EncryptionHint::Classic),
-            ],
-            false,
-        ))?;
-        let serialized_policy = <Vec<u8>>::try_from(&policy)?;
-        let parsed_policy = Policy::parse_and_convert(&serialized_policy)?;
-        // check policy size
-        assert_eq!(parsed_policy.attributes().len(), policy.attributes().len());
-
-        // check order
-        let orig_ordered_dim = policy.dimensions.get("Security Level").unwrap();
-        let parsed_ordered_dim = parsed_policy.dimensions.get("Security Level").unwrap();
-        assert_eq!(
-            parsed_ordered_dim.get_attributes_name().collect::<Vec<_>>(),
-            orig_ordered_dim.get_attributes_name().collect::<Vec<_>>(),
-        );
-
-        Ok(())
-    }
-
-    #[test]
-    pub fn test_create_policy_from_spec() -> Result<(), Error> {
-        let json = r#"
-    {
-        "Security Level::<": [
-            "Protected",
-            "Confidential",
-            "Top Secret::+"
-        ],
-        "Department": [
-            "R&D",
-            "HR",
-            "MKG",
-            "FIN"
-        ]
-    }
-    "#;
-
-        let policy_json: HashMap<String, Vec<String>> = serde_json::from_str(json).unwrap();
-        let policy: Policy = policy_json.try_into()?;
-        assert_eq!(policy.dimensions.len(), 2);
-        assert!(matches!(
-            policy.dimensions.get("Security Level").unwrap(),
-            Dimension::Ordered(_)
-        ));
-        assert!(matches!(
-            policy.dimensions.get("Department").unwrap(),
-            Dimension::Unordered(_)
-        ));
-        assert_eq!(
-            policy
-                .dimensions
-                .get("Security Level")
-                .unwrap()
-                .attributes()
-                .count(),
-            3
-        );
-        assert_eq!(
-            policy
-                .get_attribute_hybridization_hint(&QualifiedAttribute::new("Department", "MKG"))
-                .unwrap(),
-            EncryptionHint::Classic
-        );
-        assert_eq!(
-            policy
-                .get_attribute_hybridization_hint(&QualifiedAttribute::new(
-                    "Security Level",
-                    "Protected"
-                ))
-                .unwrap(),
-            EncryptionHint::Classic
-        );
-        assert_eq!(
-            policy
-                .get_attribute_hybridization_hint(&QualifiedAttribute::new(
-                    "Security Level",
-                    "Top Secret"
-                ))
-                .unwrap(),
-            EncryptionHint::Hybridized
-        );
-
-        Ok(())
-    }
-}
